Move MetadataTypeResolver to separate artifact
