+++ fahrschein-example/src/main/java/org/zalando/fahrschein/example/Main.java
@@ -9,8 +9,18 @@ import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.zalando.fahrschein.*;
-import org.zalando.fahrschein.domain.*;
+import org.zalando.fahrschein.EventProcessingException;
+import org.zalando.fahrschein.ExponentialBackoffStrategy;
+import org.zalando.fahrschein.IORunnable;
+import org.zalando.fahrschein.Listener;
+import org.zalando.fahrschein.NakadiClient;
+import org.zalando.fahrschein.NoBackoffStrategy;
+import org.zalando.fahrschein.StreamParameters;
+import org.zalando.fahrschein.ZignAccessTokenProvider;
+import org.zalando.fahrschein.domain.Cursor;
+import org.zalando.fahrschein.domain.Lock;
+import org.zalando.fahrschein.domain.Partition;
+import org.zalando.fahrschein.domain.Subscription;
 import org.zalando.fahrschein.example.domain.OrderEvent;
 import org.zalando.fahrschein.example.domain.OrderCreatedEvent;
 import org.zalando.fahrschein.example.domain.OrderPaymentAcceptedEvent;
