+++ fahrschein-example/src/main/java/org/zalando/fahrschein/example/domain/OrderEvent.java
@@ -3,8 +3,8 @@ package org.zalando.fahrschein.example.domain;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.annotation.JsonTypeResolver;
-import org.zalando.fahrschein.MetadataTypeResolver;
 import org.zalando.fahrschein.domain.Event;
+import org.zalando.fahrschein.typeresolver.MetadataTypeResolver;
 
 @JsonTypeInfo(use = JsonTypeInfo.Id.CUSTOM, visible = false)
 @JsonTypeResolver(MetadataTypeResolver.class)
